added some error checks
