+++ src/chirp/cpp/Makefile
@@ -22,13 +22,13 @@ vpath %.proto $(PROTOS_PATH)
 
 all: backend_store_server service_layer_server command_line key_value_store_tests
 
-command_line: command_line.cc backend_store.pb.o backend_store.grpc.pb.o service_layer.pb.o service_layer.grpc.pb.o server_clients.o
+command_line: command_line.cc backend_store.pb.o backend_store.grpc.pb.o service_layer.pb.o service_layer.grpc.pb.o backend_store_client.o service_layer_client.o
 	$(CXX) $^ $(LDFLAGS) -o $@ -lgtest -lgflags
 
-backend_store_server: backend_store.pb.o backend_store.grpc.pb.o service_layer.pb.o service_layer.grpc.pb.o server_clients.o backend_store_server.o key_value_store.o
+backend_store_server: backend_store.pb.o backend_store.grpc.pb.o service_layer.pb.o service_layer.grpc.pb.o backend_store_client.o service_layer_client.o backend_store_server.o key_value_store.o
 	$(CXX) $^ $(LDFLAGS) -o $@ -lgtest -lgflags
 
-service_layer_server: service_layer.pb.o service_layer.grpc.pb.o backend_store.pb.o backend_store.grpc.pb.o server_clients.o service_layer_server.o
+service_layer_server: service_layer.pb.o service_layer.grpc.pb.o backend_store.pb.o backend_store.grpc.pb.o backend_store_client.o service_layer_client.o service_layer_server.o
 	$(CXX) $^ $(LDFLAGS) -o $@ -lgtest -lgflags
 
 key_value_store_tests: key_value_store_tests.cc key_value_store.o
@@ -43,4 +43,4 @@ key_value_store_tests: key_value_store_tests.cc key_value_store.o
 	$(PROTOC) -I $(PROTOS_PATH) --cpp_out=. $<
 
 clean:
-	rm -f *.o *.pb.cc *.pb.h command_line server_clients backend_store_server service_layer_server key_value_store_tests
+	rm -f *.o *.pb.cc *.pb.h command_line backend_store_server service_layer_server key_value_store_tests
