+++ src/chirp/cpp/backend_store_server.h
@@ -9,7 +9,7 @@
 #include <grpcpp/server_builder.h>
 #include <grpcpp/server_context.h>
 #include "backend_store.grpc.pb.h"
-#include "server_clients.h"
+#include "service_layer_client.h"
 #include "key_value_store.h"
 
 class KeyValueStoreServiceImpl final : public KeyValueStore::Service {
@@ -22,6 +22,9 @@ class KeyValueStoreServiceImpl final : public KeyValueStore::Service {
                     ServerReaderWriter<GetReply, GetRequest>* stream);
     Status deletekey(ServerContext* context, const DeleteRequest* request,
                     DeleteReply* reply);
+    std::vector<std::string>* DFSReplyThread(std::map<std::string, std::vector<std::string> > chirp_map,
+                                             std::vector<std::string> *reply_thread_vec,
+                                             std::string chirp_id);
   private:
     KeyValueStoreClass store_;
 };
